+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## [0.11.16] - 2019-03-26
 - Updated activity status automatically every second
+- Updated unit tests for sortActivities
 
 ## [0.11.15] - 2019-03-25
 - Updated getEvents with changed 'users' field